Used NULL instead of 0 to represent null pointers. Use size_t where appropriate. Conf.h now allows users to externally override the header files in the private include path. All type information previously found in conf.h has been moved to private/types.h.

+++ include/bond/allocator.h
@@ -1,6 +1,8 @@
 #ifndef BOND_ALLOCATOR_H
 #define BOND_ALLOCATOR_H
 
+#include "bond/conf.h"
+
 namespace Bond
 {
 
@@ -8,11 +10,11 @@ class Allocator
 {
 public:
 	virtual ~Allocator() {}
-	virtual void *Allocate(int size) = 0;
+	virtual void *Allocate(size_t size) = 0;
 	virtual void Free(void *buffer) = 0;
 
 	template <typename T> T *Alloc() { return static_cast<T *>(Allocate(sizeof(T))); }
-	template <typename T> T *Alloc(int numElements) { return static_cast<T *>(Allocate(sizeof(T) * numElements)); }
+	template <typename T> T *Alloc(size_t numElements) { return static_cast<T *>(Allocate(sizeof(T) * numElements)); }
 };
 
 }
